+++ src/constants/filterConfigs.js
@@ -40,7 +40,13 @@ export const EBSCO_PROVIDER_ID = 19;
 
 export const selectionStatusDefaultFilterOption = selectionStatusFilterOptions.ALL;
 
+export const FILTER_TYPES = {
+  SELECT: 'select',
+  CHECKBOX: 'checkbox',
+};
+
 export const selectionStatusFilterConfig = {
+  type: FILTER_TYPES.CHECKBOX,
   name: 'selected',
   label: <FormattedMessage id="ui-eholdings.label.selectionStatus" />,
   defaultValue: selectionStatusDefaultFilterOption,
@@ -61,6 +67,31 @@ export const selectionStatusFilterConfig = {
 };
 
 export const publicationTypeFilterConfig = {
+  type: FILTER_TYPES.CHECKBOX,
+  name: 'type',
+  label: <FormattedMessage id="ui-eholdings.label.publicationType" />,
+  defaultValue: 'all',
+  options: [
+    { label: <FormattedMessage id="ui-eholdings.filter.all" />, value: 'all' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.audiobook" />, value: 'audiobook' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.book" />, value: 'book' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.book_series" />, value: 'bookseries' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.database" />, value: 'database' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.journal" />, value: 'journal' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.newsletter" />, value: 'newsletter' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.newspaper" />, value: 'newspaper' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.proceedings" />, value: 'proceedings' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.report" />, value: 'report' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streaming_audio" />, value: 'streamingaudio' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.streaming_video" />, value: 'streamingvideo' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.thesis_dissertation" />, value: 'thesisdissertation' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.website" />, value: 'website' },
+    { label: <FormattedMessage id="ui-eholdings.filter.pubType.unspecified" />, value: 'unspecified' }
+  ]
+};
+
+export const publicationTypeTitlesListFilterConfig = {
+  type: FILTER_TYPES.SELECT,
   name: 'type',
   label: <FormattedMessage id="ui-eholdings.label.publicationType" />,
   defaultValue: 'all',
@@ -84,6 +115,7 @@ export const publicationTypeFilterConfig = {
 };
 
 export const contentTypeFilterConfig = {
+  type: FILTER_TYPES.CHECKBOX,
   name: 'type',
   label: <FormattedMessage id="ui-eholdings.package.contentType" />,
   defaultValue: 'all',